% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ct_anno.R
\name{add_ct_anno}
\alias{add_ct_anno}
\title{Add cell-type level annotation
The stored elements is in meta data as a data frame (named 'ct_anno').}
\usage{
add_ct_anno(data_obj, ct_anno, match_col = NULL)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{ct_anno}{A data frame, tibble or a vector to annotate features.}

\item{match_col}{To specify, only when a tibble or a data frame input is given, which column should match that of the feature names?}
}
\value{
A SincleCellExperiment object,
}
\description{
Add cell-type level annotation
The stored elements is in meta data as a data frame (named 'ct_anno').
}
