% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_anno.R
\name{add_gene_anno}
\alias{add_gene_anno}
\title{Add gene-level statistics or annotation to the SingleCellExperiment object
The stored elements is in meta data as a data frame (named 'gene_anno').}
\usage{
add_gene_anno(data_obj, gene_anno, match_col = NULL)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{gene_anno}{A data frame, tibble or a vector to annotate features.}

\item{match_col}{To specify, only when a tibble or a data frame input is given, which column should match that of the feature names?}
}
\value{
A SincleCellExperiment object,
}
\description{
Add gene-level statistics or annotation to the SingleCellExperiment object
The stored elements is in meta data as a data frame (named 'gene_anno').
}
