% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_glob_stats.R
\name{add_glob_stats}
\alias{add_glob_stats}
\title{Add some specific global statistics to the metadata 'gene_anno' slot. Some predefined statistics are:
'det_cell_num': number of cells that express such a gene
'ave_exp_ct': mean value of a gene's average expression across all cell types
'max_exp_ct': maximum value of a gene's average expression across all cell types
'ave_rat_ct': mean value of the ratio of cells that express the gene across all cell types
'max_rat_ct': maximum value of the ratio of cells that express the gene across all cell types
'ave_exp_all': mean expression value of a gene across all cells (the data set)}
\usage{
add_glob_stats(data_obj, stats = NULL)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{stats}{A vector that select the predefined statistics for the current data set.}
}
\value{
A SingleCellExperiment object or a tibble. With meta data slots are translate into the
}
\description{
Add some specific global statistics to the metadata 'gene_anno' slot. Some predefined statistics are:
'det_cell_num': number of cells that express such a gene
'ave_exp_ct': mean value of a gene's average expression across all cell types
'max_exp_ct': maximum value of a gene's average expression across all cell types
'ave_rat_ct': mean value of the ratio of cells that express the gene across all cell types
'max_rat_ct': maximum value of the ratio of cells that express the gene across all cell types
'ave_exp_all': mean expression value of a gene across all cells (the data set)
}
