% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ct_asso.R
\name{cal_ct_asso}
\alias{cal_ct_asso}
\title{Calculate association between a trait (disease) and a cell type using specificity score after computing specificity score.
Genes with too low expression value or only minimum expression in a few cells in the data set should be filter out.
To specify how to filter genes based on these gene annotation, please use a dplyr::filter expression in string.}
\usage{
cal_ct_asso(
  data_obj,
  gene_zscore_df,
  gene_filter_setting = NULL,
  asso_model = "linear"
)
}
\arguments{
\item{data_obj}{A SingleCellExperiment object}

\item{gene_zscore_df}{A data frame containing human entrez id and trait MAGMA z-score. By default they are the first two columns.}

\item{gene_filter_setting}{A dplyr::filter expression string for gene filter across the "gene_info" slot in the object meta data. Not setting this indicating that all genes are going to be used.}

\item{asso_model}{A linear model or spearman correlation ("linear" or "spearman") to test the association between cell type specificity score and MAGMA z-score.}
}
\value{
A data frame (in metadata(data_obj)[\link{"association"}] slot) containing association p_value, FDR.
}
\description{
Calculate association between a trait (disease) and a cell type using specificity score after computing specificity score.
Genes with too low expression value or only minimum expression in a few cells in the data set should be filter out.
To specify how to filter genes based on these gene annotation, please use a dplyr::filter expression in string.
}
