% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sscore.R
\name{cal_sscore}
\alias{cal_sscore}
\title{Compute group level specificity score for genes.}
\usage{
cal_sscore(data_obj)
}
\arguments{
\item{data_obj}{SingleCellExperiment object or a tibble from cal_stat() function. The meta data contain the basic statistics for specificity score computation.}
}
\value{
A SingleCellExperiment object or a tibble, depending on the input type
}
\description{
Compute group level specificity score for genes.
}
