% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_specificity.R
\name{calc_specificity}
\alias{calc_specificity}
\title{Compute seismic specificity score for each gene and cell type.}
\usage{
calc_specificity(
  sce,
  assay_name = "logcounts",
  ct_label_col = "idents",
  min_uniq_ct = 2,
  min_ct_size = 20,
  min_cells_gene_exp = 10,
  min_avg_exp_ct = 0.1
)
}
\arguments{
\item{sce}{A SingleCellExperiment object. This object needs to include the
assay_name specified and a ct_label_col column corresponding to cell type labels
for the granularity of interest. Row names are used as the gene name identifiers.}

\item{assay_name}{An assay in sce, default 'logcounts' (calculated using scran).}

\item{ct_label_col}{A column name in colData of sce (information for the cells in sce)
that provides labels for each cell, indicating membership in each cell type
at the granularity of interest.}

\item{min_uniq_ct}{The minimum number of unique cell types needed in the data to
proceed with specificity score calculations. Default: 2.}

\item{min_ct_size}{Filter for cell types to be included in the specificity score calculation
based on the number of cells that are labeled as that cell type. Default: 20.}

\item{min_cells_gene_exp}{Filter for genes to be included in the specificity score calculation
based on the number of cells where the gene has non-zero expression. Default: 10.}

\item{min_avg_exp_ct}{Filter for genes to be included in the specificity score calculation
based on the mean of cell-type-averaged expressions (aka mean of mean cell type expression). Default: 0.1}
}
\value{
A dgeMatrix of specificity scores (genes as rows, cell types as columns).
}
\description{
Compute seismic specificity score for each gene and cell type.
}
