% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_file_helpers.R
\name{check_overlap}
\alias{check_overlap}
\title{Check for overlapping genes between specificity scores and trait gene
summaries from MAGMA.}
\usage{
check_overlap(
  sscore,
  magma,
  magma_gene_col,
  warn_thresh = 0.8,
  min_intersect = 1
)
}
\arguments{
\item{sscore}{A dgeMatrix of seismic specificity scores where
each column is a cell type and row names are gene identifiers.
(Note: the identifiers used should match those used in the MAGMA input)}

\item{magma}{A data.frame or file path to MAGMA output for a particular GWAS
with at least 2 columns: gene identifiers and z-scores.}

\item{magma_gene_col}{A character string corresponding to the column name
in the MAGMA data containing gene identifiers. Defaults to 'GENE'
from the MAGMA output.}

\item{warn_thresh}{A floating point number corresponding to the minimum
percentage of overlapping genes before the warning error is displayed.
Defaults to 80\% overlap.}

\item{min_intersect}{A floating point number representing the minimum number
of overlapping genes needed for further execution. Defaults to 1.}
}
\description{
Check for overlapping genes between specificity scores and trait gene
summaries from MAGMA.
}
