% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_inf_measure.R
\name{gene_inf_measure}
\alias{gene_inf_measure}
\title{Detect the influential genes in the a trait-cell type association pair using DFBETAS}
\usage{
gene_inf_measure(data_obj, cell_type, trait_name, pos_genes = T)
}
\arguments{
\item{data_obj}{A SingleCellExperiment object}

\item{cell_type}{The cell type.}

\item{trait_name}{The name of the trait.}

\item{pos_genes}{If only genes with positive DFBETAS value will be pick as the top influential ones?}
}
\value{
A data frame, containing gene DFBETAS value, cell ttpe gene specificity score, and trait gene MAGMA Z-score
}
\description{
Can only be used if the specified model is "linear model" and the trait-cell type association has been computed.
}
