% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEMetadataInterface.R
\name{get_ct_asso}
\alias{get_ct_asso}
\title{Get the data frame (s) of cell type-level association created by seismicGWAS.}
\usage{
get_ct_asso(data_obj, trait_name, asso_model, merge_output = FALSE)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{trait_name}{Name of the trait to get the association values. It could also be a list of traits or "all".}

\item{asso_model}{The model previously specified too calculate the cell type level association with the trait.}

\item{merge_output}{Default is FALSE. But if you wish to get a single full data frame (instead of a list) of all P_value of a list of traits, this can be set into TRUE.}
}
\value{
the data object with set values
}
\description{
Get the data frame (s) of cell type-level association created by seismicGWAS.
}
