% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_trait_associations.R
\name{get_ct_trait_associations}
\alias{get_ct_trait_associations}
\title{Calculate cell type-trait associations}
\usage{
get_ct_trait_associations(
  sscore,
  magma,
  magma_gene_col = "GENE",
  magma_z_col = "ZSTAT"
)
}
\arguments{
\item{sscore}{A dgeMatrix of seismic specificity scores where
each column is a cell type and row names are gene identifiers.
(Note: the identifiers used should match those used in the MAGMA input)}

\item{magma}{A data.frame or file path to MAGMA output for a particular GWAS
with at least 2 columns: gene identifiers and z-scores.}

\item{magma_gene_col}{A character string corresponding to the column name
in the MAGMA data containing gene identifiers. Defaults to 'GENE'
from the MAGMA output.}

\item{magma_z_col}{A character string corresponding to the column name
in the MAGMA data containing z-scores. Defaults to 'ZSTAT'
from the MAGMA output.}
}
\value{
A data.frame of trait associations for each cell type.
}
\description{
Calculate cell type-trait associations
}
