% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEMetadataInterface.R
\name{get_seismic_ct_info}
\alias{get_seismic_ct_info}
\title{Get the cell type-level mean expression, specificity score or cell number information}
\usage{
get_seismic_ct_info(data_obj, info_slot)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{info_slot}{cell type-level information slot to retrieve in the seismic analysis, any one of "mean_mat"(gene expression mean), "var_mat"(gene expression variance),
"ratio_mat"(gene expression ratio), "cell_num"(number of cells), "sscore"(gene specificity score).}
}
\value{
the value of the the information
}
\description{
Get the cell type-level mean expression, specificity score or cell number information
}
