% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{lm_loo_t}
\alias{lm_loo_t}
\title{This function is used for model diagnostics
input: two vectors
output: leave-one-out t-statistics difference}
\usage{
lm_loo_t(zscore, metric)
}
\arguments{
\item{zscore}{A zscore vector}

\item{metric}{Value to test association}
}
\description{
This function is used for model diagnostics
input: two vectors
output: leave-one-out t-statistics difference
}
