% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{lm_pvalue}
\alias{lm_pvalue}
\title{This function use a linear regression to test if two variable are positively correlated
input: sparse matrix, margin, stats and function
output:sparse matrix.}
\usage{
lm_pvalue(zscore, metric)
}
\arguments{
\item{zscore}{A zscore vector}

\item{metric}{Value to test association}
}
\description{
This function use a linear regression to test if two variable are positively correlated
input: sparse matrix, margin, stats and function
output:sparse matrix.
}
