% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_file_helpers.R
\name{load_magma_dt}
\alias{load_magma_dt}
\title{Check magma object / file is usable and returns it as a data.table}
\usage{
load_magma_dt(magma, magma_gene_col, magma_z_col)
}
\arguments{
\item{magma}{input magma object / file}

\item{magma_gene_col}{character string corresponding to the column name
in the MAGMA data containing gene identifiers}

\item{magma_z_col}{character string corresponding to the column name
in the MAGMA data containing z-scores}
}
\value{
magma file contents as a data.table
}
\description{
Check magma object / file is usable and returns it as a data.table
}
