% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zscore.R
\name{load_zscore}
\alias{load_zscore}
\title{Load z-score file(s) from MAGMA output.}
\usage{
load_zscore(file_path, name = NULL)
}
\arguments{
\item{file_path}{Path for the MAGMA .genes.out file (or directory or a list/vector of paths)}
}
\value{
A data frame with hsa_entrez and z_stat as the two columns (indicating human gene Entrez ID and its respective MAGMA z-score).
}
\description{
Load z-score file(s) from MAGMA output (with .genes.out tails). The file path could be the path of a single file or a directory or even a list of the past file.
For multiple  z-score files, the gene z-score will be combined together in a single data frame.
}
