% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmu_hsa_mapping.R
\docType{data}
\name{mmu_hsa_mapping}
\alias{mmu_hsa_mapping}
\title{Mouse to Human Gene Mapping}
\format{
A data frame with 6 columns
\describe{
\item{mmu_symbol}{mouse gene names}
\item{mmu_ensembl}{mouse gene ensembl ID}
\item{mmu_entrez}{mouse gene entrez ID}
\item{hsa_symbol}{human gene names}
\item{hsa_ensembl}{human gene ensembl ID}
\item{hsa_entrez}{human gene entrez id}
}
}
\source{
bioMart
}
\usage{
mmu_hsa_mapping
}
\description{
A data frame containing gene id mappings between mouse (Mus musculus) genes andhuman (Homo sapiens) genes.
}
\keyword{datasets}
