% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asso.R
\name{plot_asso}
\alias{plot_asso}
\title{Plot enrichment results. Return a ggplot object.}
\usage{
plot_asso(
  data_obj,
  trait,
  asso_model,
  show_value = "FDR",
  plot_top_option = NULL,
  group = NULL,
  significance = TRUE
)
}
\arguments{
\item{data_obj}{SingleCellExperiment object that contains some associtaion information in seismicGWAS.data slot (accessible via metadata(data_obj)[\link{"seismicGWAS.data"}][\link{"association}]). It could also be a data
frame or tibble, containing at least a "cell_type" column and at least "Pvalue" ansd "FDR" columns.}

\item{trait}{Which association result you would like to plot}

\item{asso_model}{The model previously specified too calculate the cell type level association with the trait.}

\item{show_value}{Show FDR or Pvalue to plot as y-axis?}

\item{group}{Which annotation column used for grouping?}

\item{significance}{To label significance level or not? By default the FDR 0.1 is asterisk, FDR 0.05 is double asterisk, FDR 0.01 is triple asterisk}

\item{plot_top_option:}{How to choose cell types to plot? This would be a string parameter passed to dplyr::filter for the data frame containing 4 columns: cell_type, Pvalue, FDR, rank. Also you can filter by cell_type_anno}
}
\value{
A ggplot object
}
\description{
Plot enrichment results. Return a ggplot object.
}
