% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_inf.R
\name{plot_gene_inf}
\alias{plot_gene_inf}
\title{Scatter plot for influential gene testing}
\usage{
plot_gene_inf(
  dfbetas_df,
  gene_label_column = "hsa_entrez",
  num_top_gene_label = 0,
  label_pos_genes = T,
  extra_labels = NULL,
  label_repel = F,
  ...
)
}
\arguments{
\item{gene_label_column}{The column name to add}

\item{num_top_gene_label}{Number of top influential genes (based on DFBETAS values) to label. If it's 0 then no genes are labelled.}

\item{label_pos_genes}{If only genes with positive Z score should be labelled? By default this is TRUE.}

\item{extra_labels}{Extra labels (besides these indicating using "num_top_gene_label" or "label_pos_genes")}

\item{label_repel}{If the labels will be expelled using the ggrepel. By default this is FALSE. This is recommended only when you have ggrepel installed.}

\item{...}{Additional arguments passed on to \code{geom_text()} or \code{geom_text_repel()}. This can include any aesthetic properties or parameters supported by \code{geom_text()}, such as \code{size}, \code{color}, \code{angle}, and so on.}

\item{dbetas_df}{The data frame from previous gene_inf_measure() function. Extra columns can be added to specify the gene annotation in the plot}
}
\value{
A ggplot object
}
\description{
Can only be used if the specified model is "linear model" and the trait-cell type association has been computed.
}
