% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_associations.R
\name{plot_top_associations}
\alias{plot_top_associations}
\title{Plot the top cell type associations for a given trait.}
\usage{
plot_top_associations(trait_res, fdr = T, limit = 10)
}
\arguments{
\item{trait_res}{A data.frame or data.table object of cell type associations
detected by seismic. It is expected that this data object contains a minimum
of three columns: cell_type, p_value, and FDR.}

\item{fdr}{A Boolean value determining if FDR should be plotted. If false
p-values are used for the plot. Defaults to printing FDR.}

\item{limit}{A positive integer limiting the number of top cell types to
be shown. Default is 10 cell types.}
}
\value{
A ggplot bar plot object.
}
\description{
Plot the top cell type associations for a given trait.
}
