% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_exp_tbl.R
\name{print_exp_tbl}
\alias{print_exp_tbl}
\title{Print expression tables for MAGMA-specificity or FUMA model - which can be taken by MAGMA.}
\usage{
print_exp_tbl(data_obj, table_type, main_table_path, aux_table_path = NULL)
}
\arguments{
\item{data_obj}{SingleCellExperiment object. Note that MAGMA-specificity takes CPM assay and FUMA takes logCPM (or logcounts) as input.}

\item{table_type}{specify either "MAGMA" or "FUMA" data tables to be printed.}

\item{main_table_path}{path of the main table.}

\item{aux_table_path}{Path of the auxiliary table. If this is not specified, not auxiliary tables will be printed out.}
}
\description{
This function will print out expression table for FUMA and MAGMA analysis. Typically the output will including two tables: one main table including the expression gene set (MAGMA) or
expression covriate file (FUMA), which could be taken by MAGMA software for cell type enrichment. The cell type are recoded in to "cluster.1, ..., cluster.n"-like format since MAGMA software
may ignore name separator or extreme long string (the order is the same to the cell type presents in the "cell_num" slot of the "group_info" metadata). The auxiliary table contains the mapping from the recoded cell type names to the real cell type names.
}
