% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_magma}
\alias{ra_magma}
\title{MAGMA genes.out output for rheumatoid arthritis}
\format{
\subsection{\code{ra_magma}}{

A data.frame with 17,460 rows and 9 columns:
\describe{
\item{GENE}{entrez gene ID}
\item{CHR}{the chromosome the gene is on}
\item{START}{start boundary of the gene on the chromosome (including any provided window during annotation)}
\item{END}{end boundary of the gene on the chromosome (including any provided window during annotation)}
\item{NSNPS}{number of SNPS annotated to gene (not excluded based on internal SNP QC)}
\item{NPARAM}{number of relevant parameters used in the MAGMA model}
\item{N}{sample size used in analyzing the gene}
\item{ZSTAT}{z-value of gene based on permuted p-value (what we also term MAGMA z-score)}
\item{P}{gene p-value}
}
}
}
\source{
\url{https://www.nature.com/articles/nature12873}
}
\usage{
ra_magma
}
\description{
A data.frame of the \code{genes.out} output from MAGMA calculated using
rheumatoid arthritis summary statistics.
}
\keyword{datasets}
