% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEMetadataInterface.R
\name{reset_seismic_analysis}
\alias{reset_seismic_analysis}
\title{Print out the summary information so far for the processing of seismicGWAS}
\usage{
reset_seismic_analysis(data_obj)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}
}
\value{
The SingleCellExperiment object with nothing in the metadata slot
}
\description{
Print out the summary information so far for the processing of seismicGWAS
}
