% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetfunctions.R
\docType{package}
\name{seismic}
\alias{seismic}
\title{Predicting information cascade by self-exciting point process model}
\description{
This package implements a self-exciting point process model for information cascades.
An information cascade occurs when many people engage in the same acts after observing
the actions of others. Typical examples are post/photo resharings on Facebook and retweets
on Twitter. The package provides functions to estimate the infectiousness of an
information cascade and predict its popularity given the observed history.
For more information, see
\url{http://snap.stanford.edu/seismic/}.
}
\references{
SEISMIC: A Self-Exciting Point Process Model for Predicting Tweet Popularity by Q. Zhao, M. Erdogdu, H. He, A. Rajaraman, J. Leskovec, ACM SIGKDD Conference on Knowledge Discovery and Data Mining (KDD), 2015.
}
