% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEMetadataInterface.R
\name{seismic_summary_info}
\alias{seismic_summary_info}
\title{Print out the summary information so far for the processing of seismicGWAS}
\usage{
seismic_summary_info(data_obj, verbose = F, info_to_return = F)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{verbose}{How detailed the information is being printed? If this is set to be False, only part of the running log will be return.
The more detailed information will be printed out if this is set to True.}

\item{info_to_return}{If true, the whole metadata}
}
\value{
The returned value depends on the input. If info_to_print parameter is null, there will be no values or data returned.
}
\description{
Print out the summary information so far for the processing of seismicGWAS
}
