% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismicGWAS-package.R
\docType{package}
\name{seismicGWAS-package}
\alias{seismicGWAS}
\alias{seismicGWAS-package}
\title{seismicGWAS: Single-cell Expression Investigation for Study of Molecular Interactions and Connections}
\description{
Discover cell type-disease associations using single-cell RNA-seq and GWAS summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ylaboratory/seismic}
  \item Report bugs at \url{https://github.com/ylaboratory/seismic/issues}
}

}
\author{
\strong{Maintainer}: Qiliang Lai \email{ql29@rice.edu}

Authors:
\itemize{
  \item Ruth Dannenfelser \email{rtdn@rice.edu} (\href{https://orcid.org/0000-0002-8766-6424}{ORCID})
  \item Vicky Yao \email{vy@rice.edu} (\href{https://orcid.org/0000-0002-3201-9983}{ORCID})
}

}
\keyword{internal}
