% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_sparse.R
\name{sweep_sparse}
\alias{sweep_sparse}
\title{Similar to the base sweep function, but can be used for sparse matrices as well.}
\usage{
sweep_sparse(x, margin, stats, fun = "*")
}
\arguments{
\item{x}{A matrix (matrix, dgeMatrix, dgTMatrix, dgCMatrix, or dgRMatrix).}

\item{margin}{Across columns or rows (margin=1 and 2 respectively)}

\item{stats}{The column name in metadata for grouping}

\item{fun}{function to be applied for each item in stats to x (along margin)}
}
\value{
A matrix
}
\description{
Enables fast computations across columns or rows (margin=1 and 2 respectively)
Only matrix, dgeMatrix, dgTMatrix, dgCMatrix, or dgRMatrix types are supported.
adapted from https://stackoverflow.com/questions/55407656/r-sweep-on-a-sparse-matrix
}
