% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmfacs_sce_small}
\alias{tmfacs_sce_small}
\title{Small Subset of Tabula Muris FACS scRNA-seq data}
\format{
\subsection{\code{tmfacs_sce_small}}{

A SingleCellExperiment with 23,341 rows (mouse gene symbols) and 4000 columns (cells),
with relevant items:
\describe{
\item{assays$logcounts}{log normalized counts calculated from scran}
\item{cluster_name}{tissue and cell type (separated by .)}
}
}
}
\source{
\url{https://tabula-muris.ds.czbiohub.org}
}
\usage{
tmfacs_sce_small
}
\description{
A subset of 4000 cells from the Tabula Muris FACS scRNA-seq data
that has already been normalized using the \code{scran} package, providing
the log-normalized counts. Cluster labels are under a labeled \code{cluster_name} column.
}
\keyword{datasets}
