% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_mmu_to_hsa_stat.R
\name{trans_mmu_to_hsa_stat}
\alias{trans_mmu_to_hsa_stat}
\title{Translate metrics named after mouse gene id to human gene id
Only genes with entrez ID are kept.
Mapping can be also done between symbol, ensembl id, etc
Features with no mapping will be discarded.}
\usage{
trans_mmu_to_hsa_stat(
  data_obj,
  gene_mapping_table,
  from,
  to,
  multi_mapping = "mean",
  metadata_type = "all"
)
}
\arguments{
\item{data_obj}{SingleCellExperiment object}

\item{gene_mapping_table}{A data frame or tibble to represent gene mapping.}

\item{from}{The column (as in gene_mapping_table) to translate the feature names from.}

\item{to}{The column (as in gene_mapping_table) to translate the feature names to.}

\item{multi_mapping}{How to translate metrics when one feature is mapped to several? By "mean" value or "sum" value? (mean for default)}

\item{metadata_type}{Meta data slots you are translating. By default all meta data (all group-level statistics and specificity score) will be translate. You can also specify anyone if you only would like to translate it.}
}
\value{
A SingleCellExperiment object or a tibble. With meta data slots are translate into the
}
\description{
Translate metrics named after mouse gene id to human gene id
Only genes with entrez ID are kept.
Mapping can be also done between symbol, ensembl id, etc
Features with no mapping will be discarded.
}
