% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{transform_sparse}
\alias{transform_sparse}
\title{This function transform sparsematrix (using such as log functions)
input: sparse matrix, margin, stats and function
output:sparse matrix.}
\usage{
transform_sparse(x, fun = "log2")
}
\arguments{
\item{x}{A sparse matrix}

\item{fun}{A function that only takes one parameter}
}
\description{
This function transform sparsematrix (using such as log functions)
input: sparse matrix, margin, stats and function
output:sparse matrix.
}
