% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_gene_ids.R
\name{translate_gene_ids}
\alias{translate_gene_ids}
\title{Translate gene ids of a given dgeMatrix of seismic specificity scores from a given organism (e.g., mouse)
to another (e.g., human) based on orthology.}
\usage{
translate_gene_ids(
  sscore,
  from = "mmu_ensembl",
  to = "hsa_entrez",
  multi_mapping = "mean"
)
}
\arguments{
\item{sscore}{dgeMatrix of specificity scores (direct output of calc_specificity())}

\item{from}{The organism and ID type of current identifiers, separated by an underscore (default: mmu_ensembl).}

\item{to}{The organism and ID type of target identifiers, separated by an underscore (default: hsa_entrez).}

\item{multi_mapping}{Method to translate metrics when one source ID has several target ID matches. Options: mean (default), sum.}
}
\value{
A dgeMatrix that has row names using target identifiers, with
}
\description{
Currently supported organisms: hsa, mmu (human, mouse, respectively)
Currently supported IDs: symbol, ensembl, entrez
}
\details{
Unmatched rows are dropped from the matrix.
}
