% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetfunctions.R
\name{tweet}
\alias{tweet}
\title{An example information cascade}
\format{
A data frame with 15563 rows and 2 columns
}
\description{
A dataset containing all the (relative) resharing time and node degree of a tweet. The original Twitter ID is 127001313513967616.
}
\details{
\itemize{
  \item relative_time_second. resharing time in seconds
  \item number_of_followers. number of followers
}
}
