% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zscore.R
\name{load_zscore}
\alias{load_zscore}
\title{Load a z-score file from MAGMA output. The output will be saved as a two-column data frame}
\usage{
load_zscore(file_path, name = NULL)
}
\arguments{
\item{file_path}{Path for the MAGMA .genes.out files}
}
\value{
A data frame with hsa_entrez and z_stat as the two columns (indicating human gene Entrez ID and its respective MAGMA z-score).
}
\description{
Load a z-score file from MAGMA output. The output will be saved as a two-column data frame
}
