% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_exp_tbl.R
\name{print_exp_tbl}
\alias{print_exp_tbl}
\title{Print expression tables for MAGMA-specificity or FUMA model - which can be taken by MAGMA.}
\usage{
print_exp_tbl(
  data_obj,
  group,
  gene_id_from,
  table_type,
  main_table_path,
  aux_table_path = NULL,
  meta_data = NULL,
  assay_name = NULL,
  filter_rare_ct = TRUE,
  filter_thres = 20
)
}
\arguments{
\item{data_obj}{SingleCellExperiment object. Note that MAGMA-specificity takes CPM assay and FUMA takes logCPM (or logcounts) as input.}

\item{group}{The column name in metadata for grouping}

\item{gene_id_from}{For the feature names, what kind of gene id is it using? This should be one of: "mmu_symbol", "mmu_ensembl", "mmu_entrez", "hsa_symbol", "hsa_ensembl", "hsa_entrez".}

\item{table_type}{specify either "MAGMA" or "FUMA" data tables to be printed.}

\item{main_table_path}{path of the main table.}

\item{aux_table_path}{Path of the auxiliary table. If this is not specified, not auxiliary tables will be printed out.}

\item{meta_data}{Metadata dataframe. By default it will be rowData(data_obj)}

\item{assay_name}{The name of the assay to extract. Recommend CPM for MAGMA-specificity and logCPM for FUMA. If this was not specified, assay_name will be dependent on the table_type parameter.}

\item{filter_rare_ct}{If cell groups with too few cells should be filtered out. By default this is set to TRUE because these groups may be contaminants or something else.}

\item{filter_thres}{If filter_rare_ct is TRUE, what cells we are going to filter out?}
}
\description{
This function will print out expression table for FUMA and MAGMA analysis. Typically the output will including two tables: one main table including the expression gene set (MAGMA) or
expression covriate file (FUMA), which could be taken by MAGMA software for cell type enrichment. The cell type are recoded in to "cluster.1, ..., cluster.n"-like format since MAGMA software
may ignore name separator or extreme long string. The auxiliary table contains the mapping from the recoded cell type names to the real cell type names.
}
