% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{sweep_sparse}
\alias{sweep_sparse}
\title{sweep_sparse function do the same as the base sweep function, but is adapted to sparse matrix.
Fast computation across columns or rows (margin=1 and 2 respectively) for the matrix.
Only dgCMatrix and dgRMatrix are accepted.}
\usage{
sweep_sparse(x, margin, stats, fun = "*")
}
\arguments{
\item{x}{A sparse matrix.}

\item{margin}{Across columns or rows (margin=1 and 2 respectively)}

\item{fun}{The column name in metadata for grouping}
}
\value{
A sparse matrix
}
\description{
sweep_sparse function do the same as the base sweep function, but is adapted to sparse matrix.
Fast computation across columns or rows (margin=1 and 2 respectively) for the matrix.
Only dgCMatrix and dgRMatrix are accepted.
}
